Move the additional reconcile logic back to k8utils. Move config related functions to kafka/common.go to prevent import cycle.
