+++ pkg/resources/kafka/configmap.go
@@ -16,21 +16,16 @@ package kafka
 
 import (
 	"bytes"
-	"context"
 	"fmt"
-	"reflect"
 	"sort"
 	"strings"
 	"text/template"
 
 	"emperror.dev/errors"
-	"github.com/banzaicloud/kafka-operator/pkg/errorfactory"
-	"github.com/banzaicloud/kafka-operator/pkg/k8sutil"
 	"github.com/go-logr/logr"
 	"github.com/imdario/mergo"
+
 	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/banzaicloud/kafka-operator/api/v1alpha1"
 	"github.com/banzaicloud/kafka-operator/api/v1beta1"
@@ -279,114 +274,4 @@ func (r Reconciler) generateBrokerConfig(id int32, brokerConfig *v1beta1.BrokerC
 	sort.Strings(completeConfig)
 
 	return strings.Join(completeConfig, "\n")
-}
-
-func (r *Reconciler) updateBrokerStatus(log logr.Logger, desired *corev1.ConfigMap) error {
-	desiredType := reflect.TypeOf(desired)
-	current := desired.DeepCopyObject()
-
-	key, err := runtimeClient.ObjectKeyFromObject(current)
-	if err != nil {
-		return errors.WithDetails(err, "kind", desiredType)
-	}
-	log = log.WithValues("kind", desiredType, "name", key.Name)
-
-	err = r.Client.Get(context.TODO(), key, current)
-	if err != nil && !apierrors.IsNotFound(err) {
-		return errorfactory.New(errorfactory.APIFailure{}, err, "getting resource failed", "kind", desiredType, "name", key.Name)
-	}
-
-	if k8sutil.CheckIfObjectUpdated(log, desiredType, current, desired) {
-		// Only update status when configmap belongs to broker
-		if id, ok := desired.Labels["brokerId"]; ok {
-			currentConfigs := util.ParsePropertiesFormat(current.(*corev1.ConfigMap).Data["broker-config"])
-			desiredConfigs := util.ParsePropertiesFormat(desired.Data["broker-config"])
-
-			var statusErr error
-			// if only per broker configs are changed, do not trigger rolling upgrade by setting ConfigOutOfSync status
-			if shouldRefreshOnlyPerBrokerConfigs(currentConfigs, desiredConfigs, log) {
-				log.V(1).Info("setting per broker config status to out of sync")
-				statusErr = k8sutil.UpdateBrokerStatus(r.Client, []string{id}, r.KafkaCluster, v1beta1.PerBrokerConfigOutOfSync, log)
-			} else {
-				statusErr = k8sutil.UpdateBrokerStatus(r.Client, []string{id}, r.KafkaCluster, v1beta1.ConfigOutOfSync, log)
-			}
-			if statusErr != nil {
-				return errors.WrapIfWithDetails(err, "updating status for resource failed", "kind", desiredType)
-			}
-		}
-	}
-
-	return nil
-}
-
-func shouldRefreshOnlyPerBrokerConfigs(currentConfigs, desiredConfigs map[string]string, log logr.Logger) bool {
-	touchedConfigs := collectTouchedConfigs(currentConfigs, desiredConfigs, log)
-
-	if _, ok := touchedConfigs[securityProtocolMapConfigName]; ok {
-		if listenersSecurityProtocolChanged(currentConfigs, desiredConfigs) {
-			return false
-		}
-	}
-
-	for _, perBrokerConfig := range perBrokerConfigs {
-		delete(touchedConfigs, perBrokerConfig)
-	}
-
-	return len(touchedConfigs) == 0
-}
-
-// Security protocol cannot be updated for existing listener
-// a rolling upgrade should be triggered in this case
-func listenersSecurityProtocolChanged(currentConfigs, desiredConfigs map[string]string) bool {
-	// added or deleted config is ok
-	if currentConfigs[securityProtocolMapConfigName] == "" || desiredConfigs[securityProtocolMapConfigName] == "" {
-		return false
-	}
-	currentListenerProtocolMap := make(map[string]string)
-	for _, listenerConfig := range strings.Split(currentConfigs[securityProtocolMapConfigName], ",") {
-		listenerProtocol := strings.Split(listenerConfig, ":")
-		if len(listenerProtocol) != 2 {
-			continue
-		}
-		currentListenerProtocolMap[strings.TrimSpace(listenerProtocol[0])] = strings.TrimSpace(listenerProtocol[1])
-	}
-	for _, listenerConfig := range strings.Split(desiredConfigs[securityProtocolMapConfigName], ",") {
-		desiredListenerProtocol := strings.Split(listenerConfig, ":")
-		if len(desiredListenerProtocol) != 2 {
-			continue
-		}
-		if currentListenerProtocolValue, ok := currentListenerProtocolMap[strings.TrimSpace(desiredListenerProtocol[0])]; ok {
-			if currentListenerProtocolValue != strings.TrimSpace(desiredListenerProtocol[1]) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// collects are the config keys that are either added, removed or updated
-// between the current and the desired ConfigMap
-func collectTouchedConfigs(currentConfigs, desiredConfigs map[string]string, log logr.Logger) map[string]struct{} {
-	touchedConfigs := make(map[string]struct{})
-
-	currentConfigsCopy := make(map[string]string)
-	for k, v := range currentConfigs {
-		currentConfigsCopy[k] = v
-	}
-
-	for configName, desiredValue := range desiredConfigs {
-		if currentValue, ok := currentConfigsCopy[configName]; !ok || currentValue != desiredValue {
-			// new or updated config
-			touchedConfigs[configName] = struct{}{}
-		}
-		delete(currentConfigsCopy, configName)
-	}
-
-	for configName := range currentConfigsCopy {
-		// deleted config
-		touchedConfigs[configName] = struct{}{}
-	}
-
-	log.V(1).Info("configs have been changed", "configs", touchedConfigs)
-	return touchedConfigs
 }
